PyBytecode interpreter: adding method calls
Creating code-objects:
They have instructions (obviously), constants,
variable names and method argument length - this
would be created by a compiler.

Code ojects live inside stack frames (callstack)
that carry the code-object itself,
the local data stack, names and constants.

Arguments to the function is placed on the local
stack by the calling frame. When the method returns,
if any value is left in it's local stack, that value
(the first found) is returned as the result
of the method call.

Verification:
* Still prints 7, but this time via a method call.