Bytecode: added calling of subroutine
Converted primitives on the const pool
to objects (ints and code-blocks)

Added calling a method which expects its arguments
to be on the stack when called - this is wrong - stuff
might happen before it's arguments are referenced.

I need to add support for either pushing values
when calling into the functions const-pool or
or adding support for variables and their values.

It probably leaks memory like my exam-studies
but it prints 5!