Bytecode: arguments to function is now in const-pool
At the far end of the const pool the arguments to the
function lives.

Instead of pushing them on the stack - insert them
at the far end of the const pool.

Why at the far end? I dunno - my naïve assumption
is that code-generation would be easier if
they are on the far end.