Use #define:d constants opcodes
And make them available in python.
It's tempting to make the opcodes an
enum - but the default storage for
an enum in C is an int, and we're wasting
bytes - since opcodes and their arguments
probably never will be more than 256.

This is a bit more verbose but more compact
so we don't waste memory on calculating what
1 + 2 + 2 is. Neat.