Mymal: mal-only implementation! Postfix notation
So it was suggested to write a macro that turned postfix
code into prefix code (s-expressions). And
here is my naïve stab at it:

For each list move the last item to pole position
in that list. For each remainding item, if
it's a list to the same thing again - otherwise
tack on the atom to the resulting list.

Test plan:
* Atoms are returned as atoms
  (macroexpand (postfix (a) -> (a)))
* Lists are returned with the last
  element first
  (macroexpand (postfix (1 2 +) -> (+ 1 2)))
* Lists within lists are recursively
  processed
  (macroexpand (postfix ((1 2 +) 3 +) -> (+ (+ 1 2 3))