Mymal: the last circular matrix solver felt suboptimal
So I wrote an even more slower suboptimal one.

This guy walks the actual matrix (or seen as a maze)
and tracks the coordinates of each step. When
the step equals the number it stops, returns the
coordinates which are then calculated (as in the
wikipedia article where the distance is
|p1 - 0| + |p2 - 0|

Test plan:
* Test-cases for this too, baby!