Mymal: Step A almost done!
Step A so close! It works almost in self-hosting
mode with all tests except for 5 failing tests-cases
that will require quite some work - this is wrapping
lists and dicts and labmdas in to objects that
can have a meta attribute on them.

This requires to return MalList wherever
a regular list is returned. Same goes for
dicts and fuctions.

Maybe some day.

Test plan:
* All tests but 5 passes for running in
  hosting mode (make MAL_IMPL=mymal "test^mal"
  The only left are associating
  meta with lists and dicts and regular lambdas.