Bytecode: start an embryo of a scheme-compiler
Following this tutorial
http://scheme2006.cs.uchicago.edu/11-ghuloum.pdf
and I'ma going to try to make that via
a compiler in python and a vm in C.

TODO
Add more test-cases on boundaries
and overflow. Or switch ot arbitrary
precision integers.

Make compiler take a file as input
so the test-program can write test-cases
by itself.

Test plan:
* Running python schemec.py produces a
  great digit of 12345 since the program
  is compiled to print a digit.
* python test_schemec.py runs a test-program
  and verifies that the program output
  is indeed 7.