Schemec: add index into source for error-messages
Better to do this now while I'm still fresh at it.

Add indexes into the original source code to be
able to print better error messages.

This means converting all primitive values
in go to structs wrapping the primitive
value and a index into the source where
that token was found.

Test plan:
* All test-cases now converted to the use
  of source-markers.